\docType{package}
\name{BellB-package}
\alias{BellB-package}
\title{Bell polynomial evaluation}
\description{
Functions for evaluating the Bell polynomials, partial and
complete.
}
\details{
Included functions: \code{BellB} for partial or complete
Bell polynomials, \code{BellN} for Bell numbers, and
\code{stirling2} for Stirling numbers of second kind.

Note: Only evaluation for a given x-vector. Does not return
coefficients or powers.
}

