\name{BellB}
\alias{BellB}
\title{Evaluate partial and complete Bell polynomial}
\usage{
BellB(n, k, x)
}
\arguments{
  \item{n}{number of elements}

  \item{k}{number of elements to take (see details)}

  \item{x}{vector of length n-k+1.

  If k is missing, computes the complete Bell polynomial
  using summation over k=1:n}
}
\description{
Evaluates the partial and complete Bell polynomial,
depending on parameters.
}
\references{
Cvijovic, D (2011). New identities for the partial Bell
polynomials. DOI: 10.1016/j.aml.2011.03.043
}

